+++ src/ContactInformation/ContactInformationContainerWrapper.jsx
@@ -2,13 +2,13 @@ import React from "react";
 import { AuthContext } from "../App";
 import ContactInformationContainer from "./ContactInformationContainer";
 
-const ContactInformationContainerWrapper = () => {
+const ContactInformationContainerWrapper = props => {
   return (
     <div className="wrapper contact-information-container-wrapper">
       <AuthContext.Consumer>
         {auth => (
           <ContactInformationContainer
-            userId={auth.userId}
+            userId={props.userId || auth.userId}
             logout={auth.logout}
             admin={auth.admin}
           />